:recycle: split logstorage in multiple systems
